StoryBoard Update
Updated StoryBoard with all screens that were defined on project
discussion

+++ GymListHelper/ViewController.m
@@ -156,13 +156,13 @@
 }
 
 // The number of columns of data
-- (int)numberOfComponentsInPickerView:(UIPickerView *)pickerView
+- (NSInteger)numberOfComponentsInPickerView:(UIPickerView *)pickerView
 {
     return 1;
 }
 
 // The number of rows of data
-- (int)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
+- (NSInteger)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
 {
     return (PICKER_MAX-PICKER_MIN+1);
 }
@@ -170,7 +170,7 @@
  //The data to return for the row and component (column) that's being passed in
 - (NSString*)pickerView:(UIPickerView *)pickerView titleForRow:(NSInteger)row forComponent:(NSInteger)component
 {
-    return [NSString stringWithFormat:@"%d", (row+PICKER_MIN)];
+    return [NSString stringWithFormat:@"%ld", (row+PICKER_MIN)];
     //[_PickerView selectedRowInComponent:0];
 }
